Conflicts:
	debian/control
	exe/application.py
	exe/engine/config.py
	exe/engine/translate.py
	exe/webui/preferencespage.py

+++ exe/jsui/mainpage.py
@@ -99,11 +99,12 @@ class MainPage(RenderableLivePage):
         # And in the main section
         self.propertiesPage = PropertiesPage(self)
         self.authoringPage = None
-        self.previewDir = None
+        self.previewPage = None
+        self.printPage = None
         self.authoringPages = {}
         self.classificationSources = {}
 
-        G.application.resourceDir=Path(package.resourceDir)
+        G.application.resourceDir = Path(package.resourceDir)
 
         self.location_buttons = LocationButtons()
 
@@ -130,6 +131,14 @@ class MainPage(RenderableLivePage):
             self.exportWebSite(None, self.package.previewDir, stylesDir)
             self.previewPage = File(self.package.previewDir / self.package.name)
         return self.previewPage
+
+    def child_print(self, ctx):
+        if not self.package.printDir:
+            stylesDir = self.config.stylesDir / self.package.style
+            self.package.printDir = TempDirPath()
+            self.exportSinglePage(None, self.package.printDir, self.config.webDir, stylesDir, True)
+            self.printPage = File(self.package.printDir / self.package.name)
+        return self.printPage
 
     def child_taxon(self, ctx):
         """
@@ -186,12 +195,10 @@ class MainPage(RenderableLivePage):
         setUpHandler(self.handleExport,          'exportPackage')
         setUpHandler(self.handleXliffExport,     'exportXliffPackage')
         setUpHandler(self.handleQuit,            'quit')
-        setUpHandler(self.handleBrowseURL,       'browseURL')
-        setUpHandler(self.handleMergeXliffPackage, 'mergeXliffPackage')
+        setUpHandler(self.handleMergeXliffPackage,   'mergeXliffPackage')
         setUpHandler(self.handleInsertPackage,   'insertPackage')
         setUpHandler(self.handleExtractPackage,  'extractPackage')
         setUpHandler(self.outlinePane.handleSetTreeSelection, 'setTreeSelection')
-        setUpHandler(self.handleClearAndMakeTempPrintDir, 'makeTempPrintDir')
         setUpHandler(self.handleRemoveTempDir,   'removeTempDir')
         setUpHandler(self.handleTinyMCEimageChoice, 'previewTinyMCEimage')
         setUpHandler(self.handleTinyMCEmath,     'generateTinyMCEmath')
@@ -220,15 +227,20 @@ class MainPage(RenderableLivePage):
             self.webServer.monitor()
 
     def render_config(self, ctx, data):
-        config = {'lastDir': G.application.config.lastDir,
-                  'locationButtons': self.location_buttons.buttons,
-                  'lang': G.application.config.locale.split('_')[0],
-                  'showPreferences': G.application.config.showPreferencesOnStart == '1' and not G.application.preferencesShowed,
-                  'loadErrors': G.application.loadErrors,
-                  'showIdevicesGrouped': G.application.config.showIdevicesGrouped == '1',
-                  'authoringIFrameSrc': '%s/authoring?clientHandleId=%s' % (self.package.name, IClientHandle(ctx).handleId),
-                  'pathSep': os.path.sep
-                 }
+        config = {
+            'lastDir': G.application.config.lastDir,
+            'locationButtons': self.location_buttons.buttons,
+            'lang': G.application.config.locale.split('_')[0],
+            'showPreferences': G.application.config.showPreferencesOnStart == '1' and not G.application.preferencesShowed,
+            'loadErrors': G.application.loadErrors,
+            'showIdevicesGrouped': G.application.config.showIdevicesGrouped == '1',
+            'authoringIFrameSrc': '%s/authoring?clientHandleId=%s' % (
+                self.package.name, IClientHandle(ctx).handleId
+            ),
+            'pathSep': os.path.sep,
+            'server': G.application.server,
+            'user': G.application.config.configDir.basename()
+        }
         G.application.preferencesShowed = True
         G.application.loadErrors = []
         return tags.script(type="text/javascript")["var config = %s" % json.dumps(config)]
@@ -413,119 +425,6 @@ class MainPage(RenderableLivePage):
         # and finally, go ahead and remove the top-level tempdir itself:
         if int(rm_top_dir) != 0:
             os.rmdir(tempdir)
-
-
-    def get_printdir_relative2web(self, exported_dir):
-        """
-        related to the following ClearParentTempPrintDirs(), return a
-        local URL corresponding to the exported_dir
-        """
-        rel_name = exported_dir[len(G.application.tempWebDir):]
-        if sys.platform[:3] == "win":
-            rel_name = rel_name.replace('\\', '/')
-        if rel_name.startswith('/'):
-            rel_name = rel_name[1:]
-        http_relative_pathname = "http://127.0.0.1:" + str(self.config.port) \
-                                     + '/' + rel_name
-        log.debug('printdir http_relative_pathname=' + http_relative_pathname)
-        return http_relative_pathname
-
-    def ClearParentTempPrintDirs(self, client, log_dir_warnings):
-        """
-        Determine the parent temporary printing directory, and clear them 
-        if safe to do so (i.e., if not the config dir itself, for example)
-        Makes (if necessary), and clears out (if applicable) the parent 
-        temporary directory.
-        The calling handleClearAndMakeTempPrintDir() shall then make a 
-        specific print-job subdirectory.
-        """
-        #
-        # Create the parent temp print dir as hardcoded under the webdir, as:
-        #           http://temp_print_dirs
-        # (eventually may want to allow this information to be configured by
-        #  the user, stored in globals, etc.)
-        web_dirname = G.application.tempWebDir
-        under_dirname = os.path.join(web_dirname,"temp_print_dirs")
-        clear_tempdir = 0
-        dir_warnings = ""
-
-        # but first need to ensure that under_dirname itself is available; 
-        # if not, create it:
-        if cmp(under_dirname,"") != 0:
-            if os.path.exists(under_dirname):
-                if os.path.isdir(under_dirname):
-                    # Yes, this directory already exists.  
-                    # pre-clean it, keeping the clutter down:
-                    clear_tempdir = 1
-                else:
-                    dir_warnings = "WARNING: The desired Temporary Print " \
-                            + "Directory, \"" + under_dirname \
-                            + "\", already exists, but as a file!\n"
-                    if log_dir_warnings:
-                        log.warn("ClearParentTempPrintDirs(): The desired " \
-                                + "Temporary Print Directory, \"%s\", " \
-                                + "already exists, but as a file!", \
-                                under_dirname)
-                    under_dirname = web_dirname
-                    # but, we can't just put the tempdirs directly underneath
-                    # the webDir, since no server object exists for it.
-                    # So, as a quick and dirty solution, go ahead and put 
-                    # them in the images folder:
-                    under_dirname = os.path.join(under_dirname,"images")
-
-                    dir_warnings += "    RECOMMENDATION: please " \
-                            + "remove/rename this file to allow eXe easier "\
-                            + "management of its temporary print files.\n"
-                    dir_warnings += "     eXe will create the temporary " \
-                           + "printing directory directly under \"" \
-                           + under_dirname + "\" instead, but this might "\
-                           +"leave some files around after eXe terminates..."
-                    if log_dir_warnings:
-                        log.warn("    RECOMMENDATION: please remove/rename "\
-                            + "this file to allow eXe easier management of "\
-                            + "its temporary print files.")
-                        log.warn("     eXe will create the temporary " \
-                            + "printing directory directly under \"%s\" " \
-                            + "instead, but this might leave some files " \
-                            + "around after eXe terminates...", \
-                            under_dirname)
-                    # and note that we do NOT want to clear_tempdir 
-                    # on the config dir itself!!!!!
-            else:
-                os.makedirs(under_dirname)
-                # and while we could clear_tempdir on it, there's no need to.
-        if clear_tempdir : 
-            # before making this particular print job's temporary print 
-            # directory underneath the now-existing temp_print_dirs, 
-            # go ahead and clear out temp_print_dirs such that we have 
-            # AT MOST one old temporary set of print job files still existing
-            # once eXe terminates:
-            rm_topdir = "0"  
-            # note: rm_topdir is passed in as a STRING since 
-            # handleRemoveTempDir expects as such from nevow's 
-            # clientToServerEvent() call:
-            self.handleRemoveTempDir(client, under_dirname, rm_topdir)
-
-        return under_dirname, dir_warnings
-
-    def handleClearAndMakeTempPrintDir(self, client, suffix, prefix, \
-                                        callback):
-        """
-        Makes a temporary printing directory, and yup, that's pretty much it!
-        """
-
-        # First get the name of the parent temp directory, after making it 
-        # (if necessary) and clearing (if applicable):
-        log_dir_warnings = 1  
-        (under_dirname, dir_warnings) = self.ClearParentTempPrintDirs( \
-                                             client, log_dir_warnings)
-
-        # Next, go ahead and create this particular print job's temporary 
-        # directory under the parent temp directory:
-        temp_dir = mkdtemp(suffix, prefix, under_dirname) 
-
-        # Finally, pass the created temp_dir back to the expecting callback:
-        client.call(callback, temp_dir, dir_warnings)
 
     def handleTinyMCEimageChoice(self, client, tinyMCEwin, tinyMCEwin_name, \
                              tinyMCEfield, local_filename, preview_filename):
@@ -786,16 +685,8 @@ class MainPage(RenderableLivePage):
         """ 
         adding the print feature in using the same export functionality:
         """
-        if exportType == 'singlePage' or exportType == 'printSinglePage':
-            printit = 0
-            if exportType == 'printSinglePage':
-                printit = 1
-            exported_dir = self.exportSinglePage(client, filename, webDir, \
-                                                 stylesDir, printit)
-            if printit == 1 and not exported_dir is None:
-                web_printdir = self.get_printdir_relative2web(exported_dir)
-                G.application.config.browser.open(web_printdir)
-
+        if exportType == 'singlePage':
+            self.exportSinglePage(client, filename, webDir, stylesDir, False)
         elif exportType == 'webSite':
             self.exportWebSite(client, filename, stylesDir)
         elif exportType == 'csvReport':
@@ -828,62 +719,30 @@ class MainPage(RenderableLivePage):
 
     def handleQuit(self, client):
         """
-        Stops the server
+        Close client session and stops the server if necessary
         """
-        # first, go ahead and clear out any temp job files still in
-        # the temporary print directory:
-        log_dir_warnings = 0  
-        # don't warn of any issues with the directories at quit,
-        # since already warned at initial directory creation
-        (parent_temp_print_dir, dir_warnings) = \
-                self.ClearParentTempPrintDirs(client, log_dir_warnings)
-
         client.close("window.location = \"quit\";")
 
-        if len(self.clientHandleFactory.clientHandles) <= 1:
-            self.webServer.monitoring = False
-            G.application.config.configParser.set('user', 'lastDir', G.application.config.lastDir)
-            try:
-                shutil.rmtree(G.application.tempWebDir, True)
-                shutil.rmtree(G.application.resourceDir, True)
-            except:
-                log.debug('Don\'t delete temp directorys. ')
-            reactor.callLater(2, reactor.stop)
-        else:
-            log.debug("Not quiting. %d clients alive." % len(self.clientHandleFactory.clientHandles))
+        if not self.webServer.application.server:
+            # first, go ahead and clear out any temp job files still in
+            # the temporary print directory:
+            log_dir_warnings = 0
+            # don't warn of any issues with the directories at quit,
+            # since already warned at initial directory creation
+            (parent_temp_print_dir, dir_warnings) = \
+                    self.ClearParentTempPrintDirs(client, log_dir_warnings)
 
-    def handleBrowseURL(self, client, url):
-        """visit the specified URL using the system browser
-
-        if the URL contains %s, substitute the local webDir
-        if the URL contains %t, show a temp file containing NEWS and README """
-        if url.find('%t') > -1:
-            release_notes = os.path.join(G.application.tempWebDir,
-                    'Release_Notes.html')
-            f = open(release_notes, 'w')
-            f.write('''<html><head><title>eXe Release Notes</title></head>
-                <body><h1>News</h1><pre>\n''')
-            try:
-                news = open(os.path.join(self.config.webDir, 'NEWS'),
-                        'r').read()
-                readme = open(os.path.join(self.config.webDir, 'README'),
-                        'r').read()
-                f.write(news)
-                f.write('</pre><hr><h1>Read Me</h1><pre>\n')
-                f.write(readme)
-            except IOError:
-                # fail silently if we can't read either of the files
-                pass
-            f.write('</pre></body></html>')
-            f.close()
-            url = url.replace('%t', release_notes)
-        else:
-            url = url.replace('%s', self.config.webDir)
-        log.debug(u'browseURL: ' + url)
-        if hasattr(os, 'startfile'):
-            os.startfile(url)
-        else:
-            G.application.config.browser.open(url, new=True)
+            if len(self.clientHandleFactory.clientHandles) <= 1:
+                self.webServer.monitoring = False
+                G.application.config.configParser.set('user', 'lastDir', G.application.config.lastDir)
+                try:
+                    shutil.rmtree(G.application.tempWebDir, True)
+                    shutil.rmtree(G.application.resourceDir, True)
+                except:
+                    log.debug('Don\'t delete temp directorys. ')
+                reactor.callLater(2, reactor.stop)
+            else:
+                log.debug("Not quiting. %d clients alive." % len(self.clientHandleFactory.clientHandles))
 
     def handleMergeXliffPackage(self, client, filename, from_source):
         """
@@ -1013,21 +872,24 @@ class MainPage(RenderableLivePage):
             if not filename.exists():
                 filename.makedirs()
             elif not filename.isdir():
-                client.alert(_(u'Filename %s is a file, cannot replace it') % 
+                if client:
+                    client.alert(_(u'Filename %s is a file, cannot replace it') %
                              filename)
                 log.error("Couldn't export web page: "+
                           "Filename %s is a file, cannot replace it" % filename)
                 return
             else:
-                client.alert(_(u'Folder name %s already exists. '
-                                'Please choose another one or delete existing one then try again.') % filename)
-                return
+                if client:
+                    client.alert(_(u'Folder name %s already exists. '
+                                'Please choose another one or delete existing one then try again.') % filename)           
+                return 
             # Now do the export
             singlePageExport = SinglePageExport(stylesDir, filename, \
                                          imagesDir, scriptsDir, cssDir, templatesDir)
             singlePageExport.export(self.package, printFlag)
         except Exception, e:
-            client.alert(_('SAVE FAILED!\n%s') % str(e))
+            if client:
+                client.alert(_('SAVE FAILED!\n%s') % str(e))
             raise
         # Show the newly exported web site in a new window
         if not printFlag:
@@ -1216,14 +1078,15 @@ class MainPage(RenderableLivePage):
         """
         Launches an exported web site or page
         """
-        if hasattr(os, 'startfile'):
-            try:
-                os.startfile(filename)
-            except UnicodeEncodeError:
-                os.startfile(filename.encode(Path.fileSystemEncoding))
-        else:
-            filename /= 'index.html'
-            G.application.config.browser.open('file://'+filename)
+        if not G.application.server:
+            if hasattr(os, 'startfile'):
+                try:
+                    os.startfile(filename)
+                except UnicodeEncodeError:
+                    os.startfile(filename.encode(Path.fileSystemEncoding))
+            else:
+                filename /= 'index.html'
+                G.application.config.browser.open('file://'+filename)
 
     def _loadPackage(self, client, filename, newLoad=True,
                      destinationPackage=None):
